requirements.txt
PyQt5>=5.15.0
python-vlc>=3.0.0
requests>=2.25.0
Pillow>=9.0.0 